Refactor Datumaro format tests
