+++ datumaro/util/test_utils.py
@@ -12,7 +12,7 @@ import unittest.mock
 import warnings
 from enum import Enum, auto
 from glob import glob
-from typing import Collection, Optional, Union
+from typing import Any, Collection, List, Optional, Union
 
 from typing_extensions import Literal
 
@@ -144,6 +144,7 @@ def compare_datasets(
     ignored_attrs: Union[None, Literal["*"], Collection[str]] = None,
     require_media: bool = False,
     require_images: bool = False,
+    **kwargs,
 ):
     compare_categories(test, expected.categories(), actual.categories())
 
@@ -194,8 +195,11 @@ def compare_datasets(
                 test.fail("ann %s, candidates %s" % (ann_a, ann_b_matches))
             item_b.annotations.remove(ann_b)  # avoid repeats
 
+    # Check dataset info
+    test.assertEqual(expected.infos(), actual.infos())
 
-def compare_datasets_strict(test, expected: IDataset, actual: IDataset):
+
+def compare_datasets_strict(test, expected: IDataset, actual: IDataset, **kwargs):
     # Compares datasets for strong equality
 
     test.assertEqual(expected.media_type(), actual.media_type())
@@ -211,6 +215,9 @@ def compare_datasets_strict(test, expected: IDataset, actual: IDataset):
         for idx, (item_a, item_b) in enumerate(zip(e_subset, a_subset)):
             test.assertEqual(item_a, item_b, "%s:\n%s\nvs.\n%s\n" % (idx, item_a, item_b))
 
+    # Check dataset info
+    test.assertEqual(expected.infos(), actual.infos())
+
 
 def compare_datasets_3d(
     test,
@@ -218,6 +225,7 @@ def compare_datasets_3d(
     actual: IDataset,
     ignored_attrs: Union[None, Literal["*"], Collection[str]] = None,
     require_point_cloud: bool = False,
+    **kwargs,
 ):
     compare_categories(test, expected.categories(), actual.categories())
 
@@ -257,6 +265,9 @@ def compare_datasets_3d(
             if ann_b is None:
                 test.fail("ann %s, candidates %s" % (ann_a, ann_b_matches))
             item_b.annotations.remove(ann_b)  # avoid repeats
+
+    # Check dataset info
+    test.assertEqual(expected.infos(), actual.infos())
 
 
 def check_save_and_load(
@@ -362,3 +373,24 @@ def mock_tfds_data(example=None, subsets=("train",)):
 
         with unittest.mock.patch("tensorflow_datasets.core.DatasetBuilder.__init__", new_init):
             yield
+
+
+class TestCaseHelper:
+    """This class will exist until we complete the migration from unittest to pytest.
+    It is designed to mimic unittest.TestCase behaviors to minimize the migration work labor cost.
+    """
+
+    def assertTrue(self, boolean: bool, err_msg: str = ""):
+        assert boolean, err_msg
+
+    def assertFalse(self, boolean: bool, err_msg: str = ""):
+        assert not boolean, err_msg
+
+    def assertEqual(self, item1: Any, item2: Any, err_msg: str = ""):
+        assert item1 == item2, err_msg
+
+    def assertListEqual(self, list1: List[Any], list2: List[Any], err_msg: str = ""):
+        assert isinstance(list1, list) and isinstance(list2, list), err_msg
+        assert len(list1) == len(list2), err_msg
+        for item1, item2 in zip(list1, list2):
+            self.assertEqual(item1, item2, err_msg)
