+++ tests/unit/data_formats/datumaro/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (C) 2023 Intel Corporation
+#
+# SPDX-License-Identifier: MIT
